chore: organize endpoints in handlers
